+++ content/changelog.md
@@ -7,6 +7,7 @@ draft: false
 ## 0.9.8 - 15 July 2026
 
 - Added 1.29 Road to Badlands types.xml entries to the [Types Explorer](https://samsobjectfinder.com/?types=1) for new objects and updated the types changelog.
+- Added Download Latest Types buttons to the Export menu and Types Explorer.
 
 
 ## 0.9.7 - 1 July 2026